refactor(ICache): move bundles to Bundles.scala
